Created convex hull, and added hash table lookup
Added sample data and created a convex hull in 3D.
The computed mesh is in HACD triangle data type which explicitly stores
its vertices. This data-type is inefficient and  needs to be converted
to  an "index-based" triangle data structure where the triangle element
is defined by indices of vertices(from a pool of vertices) that make up
that triangle.